refactor: extract core SQLite classes from electron/main.js into electron/core.js
Moves AsyncStatement, AsyncDatabase, and SQLite classes into a new
electron/core.js module with no Electron dependencies. The SQLite
constructor now accepts dbLocation as a parameter instead of calling
app.getPath('userData').

electron/main.js becomes a thin IPC wiring layer that imports from core.

This allows the core classes to be used directly in Electron utility
processes (which don't have access to the app module) without
duplicating the SQLite logic.
